- add sine rune tests - fix multiple bugs - mark requestCapabilitySetParam as optional host func for now - it might not exist e.g. for Sine

+++ include/rune_vm/DataView.hpp
@@ -8,7 +8,7 @@
 #include <cstdint>
 
 namespace rune_vm {
-    template<typename T, typename TSize = uint64_t>
+    template<typename T, typename TSize = uint32_t>
     struct DataView {
         DataView(T* data, TSize size)
             : m_data(data)
